pyproject.toml
[project]
name = "gemmology-plugin"
version = "1.1.0"
description = "Claude Code plugin for coloured gemstone expertise and crystal visualization"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
authors = [
    { name = "Fabian Schuh", email = "fabian@gemmology.dev" }
]
keywords = [
    "gemmology",
    "crystallography",
    "gemstones",
    "visualization",
    "claude-code",
    "plugin"
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering",
]
dependencies = [
    "gemmology-cdl-parser>=1.0.0",
    "gemmology-mineral-database>=1.0.0",
    "gemmology-crystal-geometry>=1.0.3",
    "gemmology-crystal-renderer>=1.0.0",
]

[project.urls]
Homepage = "https://gemmology.dev"
Documentation = "https://gemmology.dev/docs"
Repository = "https://github.com/gemmology-dev/gemmology-plugin"
Issues = "https://github.com/gemmology-dev/gemmology-plugin/issues"

[project.scripts]
gemmology = "gemmology_plugin.cli:main"
crystal-svg = "gemmology_plugin.cli:crystal_svg"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.1.0",
    "mypy>=1.0"
]
lsp = [
    "gemmology-cdl-lsp>=1.0.0"
]
all = [
    "gemmology-cdl-lsp>=1.0.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
gemmology_plugin = ["py.typed"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "C4", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
strict = false  # Integration layer with evolving APIs
ignore_missing_imports = true
check_untyped_defs = true
disable_error_code = ["attr-defined", "call-arg", "arg-type", "assignment", "type-arg"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
